+++ src/RoadRegistry.BackOffice.Handlers/RoadSegments/TranslateChangeRoadSegmentAttributesRequestHandler.cs
@@ -17,20 +17,22 @@ namespace RoadRegistry.BackOffice.Handlers.RoadSegments
 
         public TranslateChangeRoadSegmentAttributesRequestHandler()
         {
-            
+
         }
 
         public Task<ChangeRoadSegmentAttributesRequest> Handle(TranslateChangeRoadSegmentAttributesRequest request, CancellationToken cancellationToken)
         {
 
-            var roadSegmentIdentifiers = parameters.SelectMany(s => s.Wegsegmenten).Distinct().ToList();
+            //var roadSegmentIdentifiers = request.ChangeRequests.SelectMany(s => s.Wegsegmenten).Distinct().ToList();
 
-            var request = new ChangeRoadSegmentAttributesRequest();
+            //var request = new ChangeRoadSegmentAttributesRequest();
 
-            foreach (var roadSegmentId in roadSegmentIdentifiers)
-            {
-                //request.ChangeRequests.Append(new ChangeRoadSegmentStatusAttributeRequest(new RoadSegmentId(roadSegmentId)))
-            }
+            //foreach (var roadSegmentId in roadSegmentIdentifiers)
+            //{
+            //    //request.ChangeRequests.Append(new ChangeRoadSegmentStatusAttributeRequest(new RoadSegmentId(roadSegmentId)))
+            //}
+
+            throw new NotImplementedException();
         }
     }
 }