+++ src/RoadRegistry.BackOffice.Handlers/RoadSegments/UnlinkStreetNameRequestValidator.cs
@@ -4,9 +4,26 @@ using Abstractions.RoadSegments;
 using Abstractions.Validation;
 using Extensions;
 using FluentValidation;
+using FluentValidation.Results;
 
 public class UnlinkStreetNameRequestValidator : AbstractValidator<UnlinkStreetNameRequest>
 {
+    protected override bool PreValidate(ValidationContext<UnlinkStreetNameRequest> context, ValidationResult result)
+    {
+        if (context.InstanceToValidate.LinkerstraatnaamId.GetIdentifierFromPuri() <= 0
+            && context.InstanceToValidate.RechterstraatnaamId.GetIdentifierFromPuri() <= 0)
+        {
+            context.AddFailure(new ValidationFailure("request", ValidationErrors.Common.JsonInvalid.Message)
+            {
+                ErrorCode = ValidationErrors.Common.JsonInvalid.Code
+            });
+
+            return false;
+        }
+
+        return base.PreValidate(context, result);
+    }
+
     public UnlinkStreetNameRequestValidator()
     {
         RuleFor(x => x.WegsegmentId)